handle error messages onlogin, updatePassword , onSignin + first ios deploy + add connectivity plugin

+++ workmanager/lib/models/planning.dart
@@ -19,7 +19,7 @@ class Day{
   String documentId;
   DateTime startHour;  // jour et heure de début fixée
   DateTime endHour;  // jour et heure de fin fixée
-  int responsibleHour; // nombre d'heures responsable dans la journée
+  double responsibleHour; // nombre d'heures responsable dans la journée
   bool startValidated; // si le début de séance a été validé ( la séance a été crée)
   bool endValidated; // si la fin de séance a été validée ( la séance a été crée)
   String QR;         // QR code associé à la journée
